[SPARK-50641][SQL] Move `GetJsonObjectEvaluator` to `JsonExpressionEvalUtils`
